Match tex1GetFrame texId flag-test cmpwi sign (97.92%->98.25%)

+++ src/main/pi_dolphin.c
@@ -4009,11 +4009,11 @@ void tex1GetFrame(u32 texId, int unused, int* outA, int* outB, int count, u8* fr
         OSRestoreInterrupts(s);
         f46c = lbl_8035F3E8[0x21];
         f518 = lbl_8035F3E8[0x4c];
-        if ((texId & 0x80000000) != 0 && (flags & 0x2000) == 0)
+        if (((int)texId & 0x80000000) != 0 && (flags & 0x2000) == 0)
         {
             idx = 0x4b;
         }
-        else if ((texId & 0x40000000) != 0 && (flags & 0x1000) == 0)
+        else if (((int)texId & 0x40000000) != 0 && (flags & 0x1000) == 0)
         {
             idx = 0x20;
         }
